fix(db): increase size of columns holding high-precision data  (#190)
* refactor(api): update prometheus api metrics on telemetry middleware

* chore(test): sort deps to keep lint happy

* chore(api): add request duration metric

* fix(api): remove metric update logic from procedure

* chore: expose `/metrics` on both web and rest api server

* chore(blob-storage-manager): track blob storage metrics

* fix(api): check response exists

* chore(web): add feedback messages counter

* refactor(web): collect default metrics on instrumentation script

* fix(api): remove prometheus metric counter update logic from procedure handler

* test: terminate prisma connection during teardown

* fix: solve PR's review comments (#188)

* test: don't display prisma logs on test mode

* fix: use `numeric` for gas and fee-related fields on db

+++ packages/db/prisma/extensions/stats.ts
@@ -139,10 +139,10 @@ export const statsExtension = Prisma.defineExtension((prisma) =>
             SELECT
               DATE_TRUNC('day', ${dateField}) as "day",
               COUNT(hash)::INT as total_blocks,
-              SUM(blob_gas_used)::BIGINT as total_blob_gas_used,
-              SUM(blob_as_calldata_gas_used)::BIGINT as total_blob_as_calldata_gas_used,
-              SUM(blob_gas_used * blob_gas_price)::BIGINT as total_blob_fee,
-              SUM(blob_as_calldata_gas_used * blob_gas_price)::BIGINT as total_blob_as_calldata_fee,
+              SUM(blob_gas_used)::DECIMAL as total_blob_gas_used,
+              SUM(blob_as_calldata_gas_used)::DECIMAL as total_blob_as_calldata_gas_used,
+              SUM(blob_gas_used * blob_gas_price)::DECIMAL as total_blob_fee,
+              SUM(blob_as_calldata_gas_used * blob_gas_price)::DECIMAL as total_blob_as_calldata_fee,
               AVG(blob_gas_used * blob_gas_price)::FLOAT as avg_blob_fee,
               AVG(blob_as_calldata_gas_used * blob_gas_price)::FLOAT as avg_blob_as_calldata_fee,
               AVG(blob_gas_price)::FLOAT as avg_blob_gas_price
